All views are JSX.
