+++ src/App.scss
@@ -1,41 +1,3 @@
-<template>
-  <Navbar :hideBackButton="$route.meta.hideBackButton"/>
-  <div class="flex-row">
-    <router-view/>
-    <GeneratorView v-if="$store.state.settings.generator"/>
-  </div>
-</template>
-
-<script lang="ts">
-import { defineComponent, defineAsyncComponent } from 'vue'
-import Navbar from '@/components/Navbar.vue'
-import { themes } from '@/themes'
-import { stringKey } from '@/store/index'
-
-export default defineComponent({
-  name: 'Sångbok',
-  components: {
-    Navbar,
-    // This loads Generator component and generator helper functions on-demand.
-    GeneratorView: defineAsyncComponent(() => import(/* webpackChunkName: "generator" */ '@/views/Generator.vue'))
-  },
-  created () {
-    // Ugly fix that updates the theme. Also done in store.
-    try {
-      const theme = JSON.parse(window.localStorage.getItem(stringKey) || '{"settings":{"theme": undefined}}').settings.theme
-      if (theme !== undefined && Object.keys(themes).indexOf(theme) !== -1) {
-        document.body.className = theme
-      } else {
-        document.body.className = 'night'
-      }
-    } catch {
-      document.body.className = 'night'
-    }
-  }
-})
-</script>
-
-<style lang="scss">
 @use "sass:math";
 @import url('https://fonts.googleapis.com/css2?family=EB+Garamond&display=swap');
 
@@ -141,4 +103,3 @@ div.titlecontainer {
     margin: auto auto;
     width: fit-content;
   }
-</style>
