+++ src/main.ts
@@ -1,5 +1,5 @@
 import { createApp } from 'vue'
-import App from './App.vue'
+import App from './App'
 import router from './router'
 import store, { key } from './store'
 import Vue3TouchEvents from 'vue3-touch-events'
