+++ src/themes/index.ts
@@ -1,4 +1,4 @@
-// Dont forget to add import statements to App.vue.
+// Dont forget to add import statements to App.scss.
 export const themes = {
   day: '🌞',
   night: '🌙',
