+++ src/views/Search.scss
@@ -1,0 +1,10 @@
+.nohits {
+    text-align: center;
+    box-shadow: unset !important;
+    border-bottom: unset !important;
+
+    &:not(:first-child) {
+      /* Hide the nohits row if there are hits. */
+      display: none;
+    }
+  }
